Refactor most of rtosc.h/cpp into other headers
