+++ src/rtosc.c
@@ -240,61 +240,6 @@ static size_t vsosc_null(const char        *address,
     return pos;
 }
 
-void rtosc_arg_val_itr_init(rtosc_arg_val_itr* itr,
-                            const rtosc_arg_val_t* av)
-{
-    itr->av = av;
-    itr->i = itr->range_i = 0;
-}
-
-const rtosc_arg_val_t* rtosc_arg_val_itr_get(const rtosc_arg_val_itr *itr,
-    rtosc_arg_val_t* buffer)
-{
-    const rtosc_arg_val_t* result;
-    if(itr->av->type == '-')
-    {
-        if(itr->av->val.r.has_delta)
-            rtosc_arg_val_range_arg(itr->av, itr->range_i, buffer);
-        else
-            *buffer = itr->av[1];
-        result = buffer;
-    }
-    else result = itr->av;
-    return result;
-}
-
-void rtosc_arg_val_itr_next(rtosc_arg_val_itr *itr)
-{
-    // increase the range index
-    if(itr->av->type == '-')
-    {
-        // increase if the limit was reached, and the limit was not infinity (=0)
-        if(++itr->range_i >= itr->av->val.r.num && itr->av->val.r.num)
-        {
-            if(itr->av->val.r.has_delta)
-            {
-                ++itr->av;
-                ++itr->i;
-            }
-            ++itr->av;
-            ++itr->i;
-            itr->range_i = 0;
-        }
-    }
-
-    // if not inside a range (or at its beginning), increase the index
-    if(!itr->range_i)
-    {
-        if(itr->av->type == 'a')
-        {
-            itr->i += itr->av->val.a.len;
-            itr->av += itr->av->val.a.len;
-        }
-        ++itr->i;
-        ++itr->av;
-    }
-}
-
 void rtosc_v2args(rtosc_arg_t* args, size_t nargs, const char* arg_str,
                   rtosc_va_list_t* ap)
 {
@@ -384,41 +329,6 @@ size_t rtosc_vmessage(char   *buffer,
     rtosc_v2args(args, nargs, arguments, &ap2);
 
     return rtosc_amessage(buffer,len,address,arguments,args);
-}
-
-size_t rtosc_avmessage(char                  *buffer,
-                       size_t                 len,
-                       const char            *address,
-                       size_t                 nargs,
-                       const rtosc_arg_val_t *args)
-{
-    rtosc_arg_val_itr itr;
-    rtosc_arg_val_itr_init(&itr, args);
-
-    int val_max;
-    {
-        // equivalent to the for loop below, in order to
-        // find out the array size
-        rtosc_arg_val_itr itr2 = itr;
-        for(val_max = 0; itr2.i < nargs; ++val_max)
-            rtosc_arg_val_itr_next(&itr2);
-    }
-
-    STACKALLOC(rtosc_arg_t, vals, val_max);
-    STACKALLOC(char, argstr,val_max+1);
-
-    int i;
-    for(i = 0; i < val_max; ++i)
-    {
-        rtosc_arg_val_t av_buffer;
-        const rtosc_arg_val_t* cur = rtosc_arg_val_itr_get(&itr, &av_buffer);
-        vals[i] = cur->val;
-        argstr[i] = cur->type;
-        rtosc_arg_val_itr_next(&itr);
-    }
-    argstr[i] = 0;
-
-    return rtosc_amessage(buffer, len, address, argstr, vals);
 }
 
 size_t rtosc_amessage(char              *buffer,
